requirements.txt
streamlit
pandas
requests
pytesseract
google-generativeai
pdf2image
Pillow
easyocr
python-dateutil